simple-crew-backend/pyproject.toml
[project]
name = "simple-crew-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools[mcp]>=1.10.1",
    "crewai[google-genai,tools]>=1.10.1",
    "fastapi[standard]>=0.135.1",
    "langchain-neo4j>=0.3.0",
    "neo4j>=5.28.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
